Moved the validation.js file
