+++ server/routes/auth.js
@@ -1,7 +1,7 @@
 const router = require("express").Router();
 const User = require("../model/User");
 const jwt = require("jsonwebtoken");
-const { registerValidation, loginValidation } = require("../../validation");
+const { registerValidation, loginValidation } = require("./validation");
 const bcrypt = require("bcryptjs");
 
 // Routes
